+++ ynr/apps/elections/api/next/serializers.py
@@ -3,8 +3,8 @@ from rest_framework import serializers
 from api.next.serializers import OrganizationSerializer
 from candidates import models as candidates_models
 from elections import models as election_models
-from official_documents.serializers import OfficialDocumentSerializer
-from popolo.serializers import (
+from official_documents.api.next.serializers import OfficialDocumentSerializer
+from popolo.api.next.serializers import (
     CandidacyOnBallotSerializer,
     MinimalPostSerializer,
 )
