+++ ynr/apps/uk_results/api/v09/serializers.py
@@ -2,7 +2,7 @@ from rest_framework import serializers
 
 from api.v09.serializers import MembershipSerializer
 
-from .models import CandidateResult, ResultSet
+from uk_results.models import CandidateResult, ResultSet
 
 
 class CandidateResultSerializer(serializers.HyperlinkedModelSerializer):
