+++ ynr/apps/uk_results/views/__init__.py
@@ -1,2 +1,1 @@
-from .api import *  # noqa
 from .base_views import *  # noqa
